% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search datasets from datos.madrid.es}
\usage{
search_datasets(keyword)
}
\arguments{
\item{keyword}{A character string to search for (e.g. "bicicleta", "eventos")}
}
\value{
A tibble data frame with dataset titles
}
\description{
Searches the Madrid open data portal for datasets matching a keyword.
}
\examples{
search_datasets("bicicleta")

}
